--- examples/ai-billing/app/api/stripe/route.ts
@@ -1,8 +1,0 @@
-export async function POST(request: Request) {
-  const data = await request.json();
-  if (data.type === "checkout.session.completed") {
-    return new Response("OK", { status: 200 });
-  }
-
-  return new Response("Bad Request", { status: 400 });
-}
