+++ examples/ai-billing/app/layout.tsx
@@ -16,8 +16,14 @@ import {
 } from "@/components/ui/dropdown-menu";
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  metadataBase: new URL("https://unkey.dev"),
+  title: "AI billing example with Unkey",
+  description:
+    "Simple AI image generation application. Contains example code of generating and refilling Unkey API keys in response to a Stripe payment link, and using the `remaining` field for measuring usage.",
+  openGraph: {
+    title: "AI billing example with Unkey",
+    images: ["https://unkey.dev/images/templates/unkey-stripe.png"],
+  },
 };
 
 export default async function RootLayout({
