+++ examples/ai-billing/package.json
@@ -30,7 +30,8 @@
     "sonner": "^1.3.1",
     "stripe": "^14.10.0",
     "tailwind-merge": "^2.2.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "unkey": "workspace:^"
   },
   "devDependencies": {
     "@types/node": "^20.10.5",
